+++ database/migrations/2025_04_07_155454_create_packages_table.php
@@ -20,8 +20,8 @@ return new class extends Migration {
             $table->double('weight_price')->default(0);
             $table->double('total_price')->default(0);
             $table->text('note')->nullable();
-            $table->unsignedInteger('status')->default(0);
-            $table->unsignedInteger('payment_status')->default(0);
+            $table->unsignedTinyInteger('status')->default(0);
+            $table->unsignedTinyInteger('payment_status')->default(0);
             $table->timestamps();
         });
     }
